# backend-spring/HealthCARE-App-mrx-part4/swagger-DOCUMENTATION.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "HealthCare+ API",
                 "description":  "Documentation officielle de l\u0027API HealthCare+ : authentification JWT, gestion des patients, médecins, rendez-vous et dossiers médicaux.",
                 "contact":  {
                                 "name":  "HealthCare+",
                                 "email":  "support@healthcare.local"
                             },
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  []
                     }
                 ],
...
    "paths":  {
...
                  "/rendezvous/{id}":  {
                                           "put":  {
                                                       "tags":  [
                                                                    "RendezVous"
                                                                ],
                                                       "operationId":  "update_1",
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "id",
                                                                              "in":  "path",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int32"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/RendezVousUpdateDTO"
                                                                                                                               }
                                                                                                                }
                                                                                       },
                                                                           "required":  true
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/RendezVousGetDTO"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 }
                                                                     },
                                                       "security":  [
                                                                        {
                                                                            "bearerAuth":  [

                                                                                           ]
                                                                        }
                                                                    ]
                                                   },
                                           "delete":  {
                                                          "tags":  [
                                                                       "RendezVous"
                                                                   ],
                                                          "operationId":  "delete_1",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "id",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int32"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK"
                                                                                    }
                                                                        },
                                                          "security":  [
                                                                           {
                                                                               "bearerAuth":  [

                                                                                              ]
                                                                           }
                                                                       ]
                                                      }
                                       },
...
                  "/rendezvous":  {
                                      "get":  {
                                                  "tags":  [
                                                               "RendezVous"
                                                           ],
                                                  "operationId":  "list_1",
                                                  "parameters":  [
                                                                     {
                                                                         "name":  "page",
                                                                         "in":  "query",
                                                                         "required":  false,
                                                                         "schema":  {
                                                                                        "type":  "integer",
                                                                                        "format":  "int32",
                                                                                        "default":  0
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "size",
                                                                         "in":  "query",
                                                                         "required":  false,
                                                                         "schema":  {
                                                                                        "type":  "integer",
                                                                                        "format":  "int32",
                                                                                        "default":  5
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "sort",
                                                                         "in":  "query",
                                                                         "required":  false,
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "default":  "dateRendezVous"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "direction",
                                                                         "in":  "query",
                                                                         "required":  false,
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "default":  "asc"
                                                                                    }
                                                                     }
                                                                 ],
                                                  "responses":  {
                                                                    "200":  {
                                                                                "description":  "OK",
                                                                                "content":  {
                                                                                                "*/*":  {
                                                                                                            "schema":  {
                                                                                                                           "$ref":  "#/components/schemas/PageRendezVousGetDTO"
                                                                                                                       }
                                                                                                        }
                                                                                            }
                                                                            }
                                                                },
                                                  "security":  [
                                                                   {
                                                                       "bearerAuth":  [

                                                                                      ]
                                                                   }
                                                               ]
                                              },
                                      "post":  {
                                                   "tags":  [
                                                                "RendezVous"
                                                            ],
                                                   "operationId":  "create_1",
                                                   "requestBody":  {
                                                                       "content":  {
                                                                                       "application/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/RendezVousPostDTO"
                                                                                                                           }
                                                                                                            }
                                                                                   },
                                                                       "required":  true
                                                                   },
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "content":  {
                                                                                                 "*/*":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/RendezVousGetDTO"
                                                                                                                        }
                                                                                                         }
                                                                                             }
                                                                             }
                                                                 },
                                                   "security":  [
                                                                    {
                                                                        "bearerAuth":  [

                                                                                       ]
                                                                    }
                                                                ]
                                               }
                                  },
...
                  "/rendezvous/search":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "RendezVous"
                                                                  ],
                                                         "operationId":  "searchByStatut",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "statut",
                                                                                "in":  "query",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string"
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "page",
                                                                                "in":  "query",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int32",
                                                                                               "default":  0
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "size",
                                                                                "in":  "query",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int32",
                                                                                               "default":  5
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "sort",
                                                                                "in":  "query",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "default":  "dateRendezVous"
                                                                                           }
                                                                            },
                                                                            {
                                                                                "name":  "direction",
                                                                                "in":  "query",
                                                                                "required":  false,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "default":  "asc"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/PageRendezVousGetDTO"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   }
                                                                       },
                                                         "security":  [
                                                                          {
                                                                              "bearerAuth":  [

                                                                                             ]
                                                                          }
                                                                      ]
                                                     }
                                         },
...
                  "/rendezvous/me":  {
                                         "get":  {
                                                     "tags":  [
                                                                  "RendezVous"
                                                              ],
                                                     "operationId":  "myRendezVous",
                                                     "parameters":  [
                                                                        {
                                                                            "name":  "page",
                                                                            "in":  "query",
                                                                            "required":  false,
                                                                            "schema":  {
                                                                                           "type":  "integer",
                                                                                           "format":  "int32",
                                                                                           "default":  0
                                                                                       }
                                                                        },
                                                                        {
                                                                            "name":  "size",
                                                                            "in":  "query",
                                                                            "required":  false,
                                                                            "schema":  {
                                                                                           "type":  "integer",
                                                                                           "format":  "int32",
                                                                                           "default":  5
                                                                                       }
                                                                        }
                                                                    ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "content":  {
                                                                                                   "*/*":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/PageRendezVousGetDTO"
                                                                                                                          }
                                                                                                           }
                                                                                               }
                                                                               }
                                                                   },
                                                     "security":  [
                                                                      {
                                                                          "bearerAuth":  [

                                                                                         ]
                                                                      }
                                                                  ]
                                                 }
                                     },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "RendezVousUpdateDTO":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "dateRendezVous":  {
                                                                                                             "type":  "string",
                                                                                                             "format":  "date-time"
                                                                                                         },
                                                                                      "statut":  {
                                                                                                     "type":  "string",
                                                                                                     "minLength":  1
                                                                                                 },
                                                                                      "patientId":  {
                                                                                                        "type":  "integer",
                                                                                                        "format":  "int32"
                                                                                                    },
                                                                                      "medcinId":  {
                                                                                                       "type":  "integer",
                                                                                                       "format":  "int32"
                                                                                                   }
                                                                                  },
                                                                   "required":  [
                                                                                    "dateRendezVous",
                                                                                    "medcinId",
                                                                                    "patientId",
                                                                                    "statut"
                                                                                ]
                                                               },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "name":  "bearerAuth",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}